webserver/requirements.txt
fastapi
uvicorn
psycopg2-binary
# remote-pdb
